start working on handling canvas interactivity
The initial work involves starting to think about the DPR and scaling the
canvas accordingly. We have to recreate the surface each time and I've
noticed a wiggle in the elements when doing the resizing.

For resizing I've used the ResizeObserver as it's well supported now and
started thinking about debouncing and throttling the events to make the
interactivity smoother.