Make graphic interfaces compatible with HTMLCanvas
This involved going back to returning an interface in the graphics and
renderer hook, as well as updating the renderer to match the attribute
setter method for setting styles.

I've seperated out the interfaces to a graphics type.

Also added the scene class to the global state, as it'll contain things
that lot's of components will likely need.