Continue working on getting panning/zooming working
This includes adding a new hand tool that'll use to test and work with
the panning/zooming.

Tweak the graphics context to init and create another surface for the
static canvas and refactor top level props to allow it.

I've also decided to store the canvas state on the scene class and added
the UPM as a global state setting.