Massive refactor after starting work on panning.
The notion of an Editor has come back into play, that controls the child
components of the things like the viewport, painter and scene.

This created a big refactor in the canvas context to only return the
canvas references and allow the editor to hold the graphics context for
the interactive and static scenes.

This has cleaned up the global store, however the cleanup will need to
continue in other places.