Have a first go at a functioning panning hand tool.
Started fleshing out the frame handler (if I'll ever need it) and added
some small methods to the painter class.

The bulk of the logic is in the hand tool and I've learnt a thing or two
about how to properly initiate the drawing pipeline. Namely, the save(),
clear(), transform() and restore().