Getting started with adding zooming in/out.
Added some simple keypress handling for the zooming in and out. Also
added some methods to the viewport for the zooming along with some
utility methods.

I also simplified surface creation and combined recreate/create so when
we call create with the surface already there we delete it and create a
new one.

Added a concrete 2d point type to clean up some of the return types in
the viewport and most probably elsewhere.