Start drawing the static metric guides
Added the ability to pan whilst holding the space button. Added a metric
for the zoom but doesn't really work well as it only updates on a
mousemove however I'm not sure I want to tie this to React state as
people can spam the zoom button.

Also added an optional alpha channel for painting.