Start implementing the cached paths in CanvasKit
The IRenderer now provides a factory method to produce the interface of
the Path that we can cache.

The CanvasKit renderer now has an internal path that will be reused as
well as a paint object to reduce WASM calls.

As an aside in the future we have to appropriately think of the mental
model of how the rendering APIs will be called for filling and stroke
shapes and whether we want to reset these on saves, which is trivial
however the restore may be a bit trickier.