Continue with perf focused refactors
Have properly integrated with the resize observer for canvas size
changes. Cleaned up the Viewport class and made the API simpler.
Added proper rAF scheduling for drawing calls. Refactored the renderer
to cache the chroma colour objects to avoid the overhead of calling it
on every redraw which is especially noticeable on mouse events. Starting
to add Editor to the tools.