Beginning redesigning how we manage the Scene
Let's rename Paths to Contours (which is how they are named in other
font projects) as to be consistent but also to not conflict with our
IPath interface that describes the actual renderable path.

Have decoupled the static guide paths generation and caching to the
Scene object as this is where all the Contour paths will be handled too
and the Painter should only be concerned with rendering.

This all sets the basis to efficiently draw some simple contours (with
handles to come) so we can then start implementing moving.