Add a simple EventEmitter class
This comes off the back of thinking through how to coordinate and
communicate changes that will allow effective caching of renderable
path objects.

We will be able signal that a contour has been changed thus
invalidating it and rebuilding it's renderable path using the emitter.
This is better than setting a flag in every branch where we invalidate a
contour.