Integrated rough handle drawing
Cleaned up a few attributes and made them function getters, as a few
style changes for the handles.

Next steps are to handle not scaling the handles by zoom as other
editors do, formalising the rendering pipeline and properly managing the
renderable paths instead of creating new ones on every re-draw.