Add logic to close paths
Modified the segment iterator to handle closed paths as well as the
contour manager to create a new path when we close the current one.

Added some messy point manipulation to keep the handles a constant size
but pan and translate correctly. We can't do this with transforms unless
we draw a path inside of the native rect method as I assume the context
scales the radius, size etc parameters so we can never get a truly
constant size.

Also moved the aggregate painter commands up to the Editor where we can
scale line widths by the zoom.