Start compositing the scene
Introduced proper drawing to the static and interactive canvases as it
became apparent this was needed to better handle the different drawing
states. This meant clearing the screen with transparent rather than
white and some z-index wizardry.

This was motivated in part by trying to implement the rectangle tool
which would benefit from drawing the rectangle stroking a rect in the
interactive canvas and then "commiting" the contour and drawing it on
the interactive canvas.