Clean up event emitter and properly handle cache paths
Start emitting domain events and respond with the appropriate handlers.
Added an invalidated flag on our representation of a Path so
a renderer (if it needs to) can properly invalidate it's path cache.
This will also setup the work needed to free those paths (as CanvasKit
is WASM based), when we want to delete contours completely.

This actually exposed a subtle bug where we were not returning the
correct EntityId for the point that closes a path (we should return the
first point's ID).