Fixed segment iterator and made progress on bezier
Now we can upgrade a line segment from the pen tool when we drag a
sufficient amount of distance. Have also started work on the cursor for
hovering of the handles when the select tool is active.

As a result the Scene now exposes more methods to get all the points
in the scene or some etc. What will happen I suspect is that we don't
expose the internal representation of a node but instead allow the
Editor to query for what it needs.