Shape now actually stores the rectangle as a contour
Doing this exposed some issues with how we were transforming from UPM to
screen and vice versa. We essentially offset the coordinates by the
editor view everytime we added points which was fine when you pass in
the clientX and clientY but was not wise if we stored the points
somewhere already offset.

Fixed the painting bug on the direction handles, and simplified the
viewport management of mouse coordinates, and we have a new app icon!