Implement rudimentary moving of points
As a result the select tool has ramped up a little bit in complexity.
Mainly, the state based types had a spruce up by attaching the
state variables to states. All rectangle normalisation has moved to a
dedicated UPMRect to make it clear that this rectangle expects UPM
coordinates and the users can expect a normalised rectangle as a result.

The state machine in select can probably be simplified and improved and
there is a bit of work to do with multiselect and bounding box driven
modifications as well thinking through the actual "commiting" of the
data model, however this will come as we continue to refine the
approach.