Handle shift selecting individual points correctly
When selecting a point whilst shift clicking it should add that
point to all the selected points, otherwise all points should be cleared
and only this one selected.

Added the ability to dash strokes, which will be used for the bounding
rectangle, along the way we squashed a couple of bugs in the renderer
with paint styles.