Start introducing hovered handle states
As a result the styles objects got a small structural rehaul, and some
useful helper methods to paint the different handles and their states as
well as determining them.

The foundation for nudging has been started too, that'll allow us to
move points by increments of 1, 10 and 100 based on key press
combinations.