Start fleshing out the API to load fonts
The idea is to have a font service and adaptors that implement a trait
that'll allow reading and writing multiple formats with loose coupling.
We start with UFO. I'm thinking that'll we'll need some generator to
build the bindings and types so we can keep in sync between the
front and backend.

Also added some cleanup scripts and general tidiness activities, such as
cleaning up the ui package.json and some linting.