Started to bring in routing
The initial home page will be the grid view of the current font if
selected and a placeholder otherwise. To achieve this react-router-dom
will handle the window navigation. The tools have been updated to have
an idle state to handle when we navigate back to the home page. The navigation was
causing the canvas to freeze if we navigated away, this was because we
were destroying the editor in the CanvasContext when we instead needed
to keep it alive.

This will drive the first portion of the UFO reading, which will be
driven by displaying a subset of the glyphs in the grid and then
selecting one and displaying the contours in the editor. It's a good way
to POC this functionality and explore the backend and frontend
communicaiton channels.

Also brought in the Quit submenu item.