Improved the toolbar
I reorganised the toolbar to make it more obvious that it's now composed
ot the tools pane and a new navigation pane. I've also formalised the
routes in the application. I've added a new font information item to the
navigation pane and a new stub page. This will set the stage for a good
exercise in extracting simple font information from a UFO and sending it
up to the frontend for display.

I've also added a new app directory to house the main entry point to App
wrapper, as well as some minimum height and width on the windows.