Make progress toward reading UFO
This branch is becoming unruly so I am just going to commit it as is and
work on it incrementally on main. This started as a crate for dealing
with unicode, using fonttools to generate data for the purporses of
grouping and creating character sets to set the stage for reading in
data and previewing it in the grid and in the editor once clicked.

We have a very experimental and rough start, with a python package that
will generate all the relevant charsets and inject them directly into ts
at build time. This will be for commonly used sets and for larger ones
will probably move down to the Rust side.

The UFO reader now actually get's all the glyphs in the read file into
memory which we can now query which returns a glyph object. Sooner or
later (better sooner), Rust should generate the ts types that are shared
which will be used in the frontend.