POC on generated ts types from Rust
The push towards hooking the backend into the application made it a good
time to start exploring generating ts types that would ideally keep the
backend and frontend syncronised in the areas we care about.

As such, the packages directory is added and a shared package created to
host the shared types. I imagine the generated character sets will go
into it's on package here as well.

This will not only keep things like the Editor class in sync across the
boundaries but also keep the tauri command output types the same in the
backend and frontend.

It's all still very experimental and will need polishing but it's a good
start.