Hot reload/rebuild on ts-rs types
To support hot reload on ts-rs generated types, a new vite plugin is
introduced that watches any crates that generate shared types and
rebuilds them in cargo and then tsc in the shared package. Since ts-rs
doesn't generate a index.ts to export the types, a small utility script
does the job.

A lot of tidy up code was neccessary to make the changes palatable.
Turbo commands have been properly configured, husky now calls pnpm
instead of bun and vitetest is used instead of bun testhusky now calls
pnpm instead of bun and vitetest is used instead of bun test. The
cleanup script has been removed in favour of each package taking care of
it's own cleanup, with turbo calling into all of these.

Overall, I'm happy with these changes and am setting good foundations
for a solid developer experience with the shared rust types, which will
hopefully keep the backend and frontend more in sync in a stress-free
way.