Draw loaded UFO contours
We can now jankily draw contours that are loaded from UFO :^). There
were a couple of gotchas with ts-rs and serde, we need to make sure if
we do things like rename = "camelCase" do it on the serde macro so it
serialises, the exported types are not enough!

Next handling clockwise and counter clockwise contours makes sense and
some refactoring the changes that made this commit possible.