Show filename in the toolpane when loading a font
Loading a font sends a font-loaded event to the frontend to
then store in the global state, and display the file name in the
toolpane.

Created a new events crate which exports payload types, we still have to
type the actual event names as well. A nice introduction to the event
system in Tauri.