Allow the glyph grid display to scroll
Overflow was removed when starting work on Shift, but let's introduce it
into the glyph grid so we can see all the glyphs in the set :^).

Fixed smooth point painting, before control points got the smooth style
instead of the corner points that have become smooth points. Have made
the getter for segments generic instead of fixed to a cubic segment.

Also, when loading files we should filter for files that we can load and
gray out the rest (TTF, OTF and UFO for now).