Generalise EventEmitter, add rudimentary undo
The EventEmitter is made more generic, and we now wrap the Tauri
EventEmitter, but pass an interface to the editor, so if we find
ourselves in a pure web enviornment we can switch the EventEmitter.

I've introduced a rudimentary UndoManager, this should recieve undo
commands off of the emitted events, as they are the commit actions. Now
that this is the Tauri event emitter, we can listen to these events on
the backend!

A small POC is demonstrates the undo capability for undoing placed
points and moving points. It's still buggy and an intial design but a
good start.

This all sets the stage to properly hook up the data models on the front
and backend.