Add undo for non nudge moves in select tool
Added a move point by method which makes undoing more intuitive than
moving points to an absolute position. Moving by a delta value
makes this change a lot easier.

The Select tool is becoming a bit unruly, bit this is expected as it's
probably one of the more complex tools in the pane.