Start the move to the EditEngine
The Select tool now gets an EditSession, which has the minimal set of
functions snd state. Now the Select tool only cares about picking what
points are selected and when to preview and apply edits, not __how__ to
apply edits.

This will encapsulate the coming complex vector editing constraints that
come with the selection/movement.