Move runtime pattern creation and detection into PatternMatcher
Created a new class to handle pattern matching behaviour and generation
now that we've introduced a new 5 sized window. To target more complex
sequences. Now the EditEngine is a lot simpler, it has no knowledge of
the RuleTable, it just uses the PatternMatcher to see if the current
selected point produces a match, to then apply the rule.